+++ src/Connection.php
@@ -130,6 +130,7 @@ class Connection extends BaseConnection
     }
 
     /**
+     * @deprecated will be removed in v10
      * @return Database|Transaction
      */
     protected function getDatabaseContext(): Database|Transaction