+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 - add support for snapshot queries (#215)
 - deprecate Connection::getDatabaseContext() and move logic to UseMutations::getMutationExecutor() (#227)
 - add support for `Query\Builder::whereNotInUnnest(...)` (#225)
-- `Query\Builder::whereIn` will now wrap values in `UNNEST` if the number of values exceeds the limit (950). (#)
+- `Query\Builder::whereIn` will now wrap values in `UNNEST` if the number of values exceeds the limit (950). (#226)
+- Commit options can now be set through config or through `Connection::setCommitOptions(...)` (#229)
 
 # v8.2.0 (2024-08-05)